requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.4.0
django-ckeditor==6.7.1
Pillow==10.4.0
gunicorn==22.0.0
whitenoise==6.7.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
